Backend/QRmaker/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.8
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
yagmail==0.15.293
yagmail==0.15.293
